project/requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
pydantic==1.7.3
starlette==0.13.6
uvicorn==0.13.3

asyncpg==0.21.0
tortoise-orm==0.16.18

pytest==6.2.0
requests==2.25.0

gunicorn==20.0.4

pytest-cov==2.10.1
flake8===3.8.4

black==20.8b1
isort==5.6.4